Новое защищённое api готово!

+++ data/API/EmailAPI/EmailResource.py
@@ -45,47 +45,45 @@ class EmailListRecourse(Resource):
 
 
 class AdminResourceEmail(Resource):
-    def get(self, email, password, email_id):
-        admin, session = check_admin_status(email, password)
-        email, session = find_by_id(email_id, session)
-        return jsonify(email.to_dict(only=('id', 'email_address')))
-
-    def delete(self, email, password, email_id):
-        admin, session = check_admin_status(email, password, 2)
-        email, session = find_by_id(email_id, session)
-        email_address = email.email_address
-        session.delete(email)
-        session.commit()
-        add_auditlog("Удаление", f"Админ {admin.name} {admin.surname} удаляет электронную почту {email_address}", admin,
-                     datetime.datetime.now())
-        return jsonify({"success": f"Электроная почта {email_address} успешно удалена"})
-
-    def put(self, email, password, email_id):
-        admin, session = check_admin_status(email, password, 2)
-        email, session = find_by_id(email_id, session)
+    def put(self, email_id):
         args, count = parser_email.parse_args(), 0
-        email_dict = email.to_dict(only=('email_address'))
-        keys = list(filter(lambda key: args[key] is not None and key in email_dict and args[key] != email_dict[key], args.keys()))
-        for key in keys:
-            count += 1
-            if key == 'email_address':
-                email.email_address = args["email_address"]
-        if count == 0:
-            return raise_error("Пустой запрос")
-        email_dict_2 = email.to_dict(only=('email_address'))
-        list_chang = [f'изменяет {key} с {email_dict[key]} на {email_dict_2[key]}' for key in keys]
-        session.commit()
-        add_auditlog("Изменение", f"Админ {admin.name} {admin.surname} изменяет электронную почту {email.email_address}:"
-                                  f" {', '.join(list_chang)}", admin, datetime.datetime.now())
-        return jsonify({"success": f"Электронная почта {email.email_address} успешно изменена"})
+        if not all(args[key] is not None for key in ['admin_email', 'action']):
+            raise_error('Пропущены некоторые важные аргументы')
+        admin, session = check_admin_status(args["admin_email"], password_manager.get_password(args["admin_email"]))
+        email, session = find_by_id(email_id, session)
+        if args['action'] == "get":
+            return jsonify(email.to_dict(only=('id', 'email_address')))
+        elif args['action'] == 'delete':
+            session.delete(email)
+            session.commit()
+            add_auditlog("Удаление", f"Админ {admin.name} {admin.surname} удаляет электронную почту {email.email_address}",
+                         admin, datetime.datetime.now())
+            return jsonify({"success": f"Электроная почта {email.email_address} успешно удалена"})
+        elif args['action'] == 'put':
+            email, session = find_by_id(email_id, session)
+            email_dict = email.to_dict(only=('email_address',))
+            keys = list(filter(lambda key: args[key] is not None and key in email_dict and args[key] != email_dict[key], args.keys()))
+            for key in keys:
+                count += 1
+                if key == 'email_address':
+                    email.email_address = args["email_address"]
+            if count == 0:
+                return raise_error("Пустой запрос")
+            email_dict_2 = email.to_dict(only=('email_address',))
+            list_chang = [f'изменяет {key} с {email_dict[key]} на {email_dict_2[key]}' for key in keys]
+            session.commit()
+            add_auditlog("Изменение", f"Админ {admin.name} {admin.surname} изменяет электронную почту {email.email_address}:"
+                                      f" {', '.join(list_chang)}", admin, datetime.datetime.now())
+            return jsonify({"success": f"Электронная почта {email.email_address} успешно изменена"})
+        raise_error("Неизвестный метод")
 
 
 class CreateEmailResource(Resource):
-    def post(self, email, password):
-        admin, session = check_admin_status(email, password)
+    def post(self):
         args = parser_email.parse_args()
-        if not all(args[key] is not None for key in ['email_address']):
+        if not all(args[key] is not None for key in ['email_address', 'admin_email']):
             raise_error('Пропущены некоторые аргументы, необходимые для создания нового адреса электронной почты')
+        admin, session = check_admin_status(args["admin_email"], password_manager.get_password(args["admin_email"]))
         if session.query(Email).filter(Email.email_address == args['email_address']).first():
             raise_error("Этот адрес электронной почты уже существует")
         new_email = Email()
